Command usage info refactored, and finding heads modularized.
1. commandInfo now represents a short usage string and a function to
perform each command. Now commands can be automatically printed with
their usage information. This was slightly complicated by an
"initialization loop", which requires the top-level 'usage' function to
close over the global 'commands' variable.

2. Searching for heads by name has been refactored into 'find*' methods
on the 'heads' type. A head can currently be in one of three
classifications: 1) active, 2) connected but disabled or 3)
disconnected.